refactor: Consolidate compliance schema loading and report generation into a single utils module
